Create and style Community component
