+++ src/constants/variant.js
@@ -1,0 +1,12 @@
+const scaleVariants = {
+  whileInView: {
+    scale: [0, 1],
+    opacity: [0, 1],
+    transition: {
+      duration: 1,
+      ease: 'easeInOut',
+    },
+  },
+};
+
+export default scaleVariants;
