+++ src/containers/index.js
@@ -2,3 +2,4 @@ export { default as Header } from './Header';
 export { default as About } from './About';
 export { default as Courses } from './Courses';
 export { default as Testimonials } from './Testimonials';
+export { default as Community } from './Community';
