+++ src/index.css
@@ -103,6 +103,20 @@
   .testimonial p {
     @apply text-xs leading-[14px] md:text-sm text-neutral;
   }
+
+  .community-section {
+    @apply max-w-full h-full px-4 py-8 text-center md:px-20 mt-24 bg-primary text-white;
+  }
+
+  .community-section h2 {
+    @apply font-clash font-bold text-lg md:text-2xl md:text-[56px] leading-[29.52px]
+      md:leading-[68px] w-3/4 md:w-3/4 capitalize mx-auto;
+  }
+
+  .community-section button {
+    @apply blue-button bg-white text-sm md:text-[16px] text-primary font-bold
+      shadow-md hover:bg-secondaryText hover:text-white;
+  }
 }
 
 @media screen and (max-width: 768px) {
